requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
attrs==25.1.0
beautifulsoup4==4.13.3
billiard==4.2.1
bleach==6.2.0
bokeh==3.6.3
cachetools==5.5.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.1.1
colorcet==3.1.0
contourpy==1.3.1
cucim-cu12==25.2.0
cuda-bindings==12.8.0
cuda-python==12.8.0
cudf-cu12==25.2.0
cugraph-cu12==25.2.0
cuml-cu12==25.2.0
cuproj-cu12==25.2.0
cupy-cuda12x==13.3.0
cuspatial-cu12==25.2.0
cuvs-cu12==25.2.0
cuxfilter-cu12==25.2.0
dask==2024.12.1
dask-cuda==25.2.0
dask-cudf-cu12==25.2.0
dask-expr==1.1.21
datashader==0.17.0
defusedxml==0.7.1
distributed==2024.12.1
distributed-ucxx-cu12==0.42.0
fastapi==0.115.8
fastjsonschema==2.21.1
fastrlock==0.8.3
filelock==3.17.0
flower==2.0.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.2.0
geopandas==1.0.1
h11==0.14.0
holoviews==1.20.1
humanize==4.12.1
idna==3.10
imageio==2.37.0
isoduration==20.11.0
Jinja2==3.1.5
joblib==1.4.2
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_proxy==4.4.0
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
kombu==5.4.2
lazy_loader==0.4
libcudf-cu12==25.2.0
libcugraph-cu12==25.2.0
libcuml-cu12==25.2.0
libcuspatial-cu12==25.2.0
libcuvs-cu12==25.2.0
libkvikio-cu12==25.2.0
libraft-cu12==25.2.0
libucx-cu12==1.18.0
libucxx-cu12==0.42.0
linkify-it-py==2.0.3
llvmlite==0.43.0
locket==1.0.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistune==3.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multipledispatch==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.4.2
numba==0.60.0
numba-cuda==0.2.0
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvcomp-cu12==4.1.0.6
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvtx==0.2.11
nx-cugraph-cu12==25.2.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
panel==1.6.1
param==2.2.0
partd==1.4.2
pillow==11.1.0
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
psutil==7.0.0
ptyprocess==0.7.0
pyarrow==19.0.1
pycparser==2.22
pyct==0.5.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pylibcudf-cu12==25.2.0
pylibcugraph-cu12==25.2.0
pylibraft-cu12==25.2.0
pynvjitlink-cu12==0.5.0
pynvml==12.0.0
pyogrio==0.10.0
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytz==2025.1
pyviz_comms==3.0.4
PyYAML==6.0.2
pyzmq==26.2.1
raft-dask-cu12==25.2.0
rapids-dask-dependency==25.2.0
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rmm-cu12==25.2.0
rpds-py==0.23.1
scikit-image==0.25.2
scipy==1.15.2
Send2Trash==1.8.3
setuptools==75.8.2
shapely==2.0.7
simpervisor==1.0.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.45.3
sympy==1.13.1
tblib==3.0.0
terminado==0.18.1
tifffile==2025.2.18
tinycss2==1.4.0
toolz==1.0.0
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
treelite==4.4.1
triton==3.2.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
uc-micro-py==1.0.3
ucx-py-cu12==0.42.0
ucxx-cu12==0.42.0
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
xarray==2025.1.2
xyzservices==2025.1.0
yarl==1.18.3
zict==3.0.0
